Move SslHandshakeFailure to separate class
